# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AICSS Backend",
    "description": "AI Cinematic Spatial System — Depth + Segmentation + Spatial Layers",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/aicss/v2/projects/{project_id}/shots/{shot_id}/archive/download": {
      "get": {
        "tags": [
          "v2 - Shots"
        ],
        "summary": "Download Shot Archive",
        "description": "下载最近一次生成的 shot archive ZIP（若不存在则现场打包）。",
        "operationId": "download_shot_archive_api_aicss_v2_projects__project_id__shots__shot_id__archive_download_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "shot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Shot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/aicss/settings/": {
      "get": {
        "tags": [
          "Settings",
          "Settings"
        ],
        "summary": "Get Settings",
        "description": "Return the current runtime settings (sensitive fields masked).",
        "operationId": "get_settings_api_aicss_settings__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Settings Api Aicss Settings  Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings",
          "Settings"
        ],
        "summary": "Post Settings",
        "description": "Apply a partial update and return the new settings snapshot.",
        "operationId": "post_settings_api_aicss_settings__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Post Settings Api Aicss Settings  Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/aicss/settings": {
      "get": {
        "tags": [
          "Settings",
          "Settings"
        ],
        "summary": "Get Settings",
        "description": "Return the current runtime settings (sensitive fields masked).",
        "operationId": "get_settings_api_aicss_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Get Settings Api Aicss Settings Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Settings",
          "Settings"
        ],
        "summary": "Post Settings",
        "description": "Apply a partial update and return the new settings snapshot.",
        "operationId": "post_settings_api_aicss_settings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Post Settings Api Aicss Settings Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SettingsUpdate": {
        "properties": {
          "model_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Mode",
            "description": "Global model mode: 'cloud' | 'local'"
          },
          "vlm_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vlm Mode",
            "description": "VLM mode: 'cloud' | 'local'"
          },
          "image_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Mode",
            "description": "Image generation mode: 'cloud' | 'local'"
          },
          "video_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Mode",
            "description": "Video generation mode: 'cloud' | 'local'"
          },
          "llm_base_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Base Url",
            "description": "OpenAI-compatible base URL for the local LLM server"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model",
            "description": "Local LLM model name (e.g. 'qwen2.5-7b-q4_k_m')"
          },
          "image_model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Model Id",
            "description": "Diffusers model ID for local image generation"
          },
          "image_dtype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Dtype",
            "description": "Dtype for image generation: 'float16' | 'bfloat16' | 'float32'"
          },
          "video_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Provider",
            "description": "Video provider: 'dashscope' | 'local_wan' | 'svd'"
          },
          "dashscope_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Api Key",
            "description": "[deprecated] Legacy single DashScope API key"
          },
          "dashscope_llm_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Llm Api Key",
            "description": "DashScope API key for LLM calls"
          },
          "dashscope_vlm_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Vlm Api Key",
            "description": "DashScope API key for VLM (vision) calls"
          },
          "dashscope_image_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Image Api Key",
            "description": "DashScope API key for image generation"
          },
          "dashscope_video_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Video Api Key",
            "description": "DashScope API key for video generation"
          },
          "dashscope_llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Llm Model",
            "description": "DashScope LLM model ID (e.g. 'qwen-plus')"
          },
          "dashscope_vlm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Vlm Model",
            "description": "DashScope VLM model ID (e.g. 'qwen-vl-chat-v1')"
          },
          "dashscope_image_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dashscope Image Model",
            "description": "DashScope image model ID (e.g. 'wanx-v1')"
          },
          "cloud_llm_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Llm Provider",
            "description": "Cloud provider for LLM: 'dashscope' | 'toapi' (legacy)"
          },
          "cloud_vlm_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Vlm Provider",
            "description": "Cloud provider for VLM: 'dashscope' | 'toapi' (legacy)"
          },
          "cloud_image_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Image Provider",
            "description": "Cloud provider for image: 'dashscope' | 'toapi' (legacy)"
          },
          "cloud_video_provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloud Video Provider",
            "description": "Cloud provider for video: 'dashscope' | 'toapi' (legacy)"
          },
          "toapi_llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toapi Llm Model",
            "description": "ToAPIs LLM model ID (legacy)"
          },
          "toapi_image_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toapi Image Model",
            "description": "ToAPIs image model ID (legacy)"
          },
          "toapi_video_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toapi Video Model",
            "description": "ToAPIs video model ID (legacy)"
          },
          "toapi_llm_api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toapi Llm Api Key",
            "description": "ToAPIs API key (legacy; shared by LLM/Image/Video)"
          },
          "providers": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providers",
            "description": "User-defined cloud provider registry"
          }
        },
        "type": "object",
        "title": "SettingsUpdate",
        "description": "Partial update payload. Any omitted field is left unchanged."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}